requirements.txt
discord.py
schedule
ics
pytz
